# Performance improvements in Sum() for Array type