@stephentoub 2020-03-11
> I don't think that the performance penalty for the single extra cast is high.

Based on what?  It is measurable.  I just ran a benchmark that compares the current implementation against your proposed implementation, using a non-int[] as the input (I used Enumerable.Range), and even for lengths up to 16, maybe 32, the impact is measurable:

|     Method | Count |      Mean |    Error |   StdDev |
|----------- |------ |----------:|---------:|---------:|
| CurrentSum |     1 |  20.02 ns | 0.179 ns | 0.158 ns |
|     NewSum |     1 |  23.03 ns | 0.271 ns | 0.240 ns |
| CurrentSum |     2 |  24.52 ns | 0.200 ns | 0.187 ns |
|     NewSum |     2 |  26.93 ns | 0.315 ns | 0.295 ns |
| CurrentSum |     4 |  31.31 ns | 0.262 ns | 0.245 ns |
|     NewSum |     4 |  34.86 ns | 0.206 ns | 0.193 ns |
| CurrentSum |     8 |  46.43 ns | 0.524 ns | 0.465 ns |
|     NewSum |     8 |  49.97 ns | 0.137 ns | 0.121 ns |
| CurrentSum |    16 |  85.26 ns | 0.465 ns | 0.435 ns |
|     NewSum |    16 |  88.58 ns | 0.154 ns | 0.144 ns |
| CurrentSum |    32 | 146.24 ns | 1.252 ns | 1.110 ns |
|     NewSum |    32 | 147.11 ns | 0.296 ns | 0.262 ns |
| CurrentSum |    64 | 266.52 ns | 1.339 ns | 1.187 ns |
|     NewSum |    64 | 266.55 ns | 0.386 ns | 0.322 ns |

So yes, the proposed implementation makes things faster for arrays while hurting every other case.  What justifies that trade-off (plus the increased code size that every generic instantiation will need to carry around)?  I just looked through a large index of source code, for all use of the `Enumerable.Sum(this IEnumerable<int>)` overload... of the ~100 call sites, only 1 was on something that might be an `int[]`... the vast majority of the rest were all on the results of other LINQ APIs, with a few additional ones on `IList<int>`, etc.

This does not look like the right trade-off to me.  If someone with a primitive array is interested in summing and getting the best possible perf, they're much better off writing the three lines of code themselves; not only will that produce the best code, it'll avoid a non-inlineable function call and a cast, and so will be faster and better achieve their goals.