@EgorBo 2020-03-11
> If someone with a primitive array is interested in summing and getting the best possible perf, they're much better off writing the three lines of code themselves;

Or use some 3rd party, e.g. LinqFaster has SIMD-accelerated Sum
https://github.com/jackmott/LinqFaster

Also the most efficient impl is probably this one https://github.com/dotnet/machinelearning/blob/c37877249a7cc326fa60cd467b56e8087dc4edf7/src/Microsoft.ML.CpuMath/AvxIntrinsics.cs#L972-L1058 🙂