@stephentoub 2020-03-11
> Or use some 3rd party
> Also the most efficient impl is probably this one

Yup.  System.Linq needs to be a good all-around solution, accomodating many varied workloads, including ones that are more sensitive to start-up time and working set than they are to throughput (e.g. Enumerable.Sum is frequently invoked on small, non-array inputs).  If you know you've got a large primitive array, you can certainly do better, complete with all the vectorization and other optimizations you can throw at it.  We're not catering to that with the built-in System.Linq.  Thankfully, for specialty needs there are specialty solutions, as @EgorBo points out.