@sonnemaf 2020-03-11
OK, I understand now what you mean with the Cast overhead. This is a valid point. Would an extra overload with an array be an option. I can't imagine but just checking.

I knew LinqFaster, never saw the ML solution. I think we can close this PR. Learned something new. Thanks for your effort.